tensorflow-api/requirements.txt
pandas==0.21.1
numpy==1.11.0
matplotlib==1.5.1
six==1.11.0
tensorflow_gpu==1.4.1
ipython==6.2.1
Pillow==5.0.0
tensorflow-gpu==1.4.1
utils==0.9.0
